requirements.txt
Flask==2.0.1
gunicorn==20.1.0
line-bot-sdk==2.0.1
pip>=21.0
setuptools>=56.0
wheel>=0.36.0
aiohttp==3.8.4
werkzeug==2.0.3
bs4==0.0.2